vibe code conversion to TS + switch to bun
