--- commands/ping.js
@@ -1,18 +1,0 @@
-const { SlashCommandBuilder } = require("discord.js");
-
-module.exports = {
-  data: new SlashCommandBuilder()
-    .setName("ping")
-    .setDescription("Replies with pong!"),
-  async execute(interaction) {
-    await interaction.reply({
-      embeds: [
-        {
-          color: 0xffc916,
-          title: "Ping?",
-          description: "Pong!",
-        },
-      ],
-    });
-  },
-};
