--- deploy.js
@@ -1,46 +1,0 @@
-const fs = require("node:fs");
-const path = require("node:path");
-const { REST, Routes } = require("discord.js");
-require("dotenv").config({ quiet: true });
-
-const commands = [];
-const commandsPath = path.join(__dirname, "commands");
-const commandFiles = fs
-  .readdirSync("./commands")
-  .filter((file) => file.endsWith(".js"));
-
-const clientId = process.env.CLIENT_ID;
-const guildId = process.env.GUILD_ID;
-
-for (const file of commandFiles) {
-  const filePath = path.join(commandsPath, file);
-  const command = require(filePath);
-  if ("data" in command && "execute" in command) {
-    commands.push(command.data.toJSON());
-  } else {
-    console.log(
-      `[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`,
-    );
-  }
-}
-
-const rest = new REST().setToken(process.env.DISCORD_BOT_TOKEN);
-
-(async () => {
-  try {
-    console.log(
-      `Started reloading ${commands.length} guild application (/) commands.`
-    );
-
-    const data = await rest.put(
-      Routes.applicationGuildCommands(clientId, guildId),
-      { body: commands }
-    );
-
-    console.log(
-      `Successfully reloaded ${data.length} guild application (/) commands.`
-    );
-  } catch (error) {
-    console.error(error);
-  }
-})();
