+++ index.ts
@@ -1,13 +1,18 @@
-// Core codes
-
-const fs = require("node:fs");
-const path = require("node:path");
-const {
+import fs from "node:fs";
+import path from "node:path";
+import { fileURLToPath, pathToFileURL } from "node:url";
+import {
   Client,
   Collection,
   GatewayIntentBits,
   Partials,
-} = require("discord.js");
+} from "discord.js";
+import dotenv from "dotenv";
+import type { Command, Event } from "./types.js";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 const client = new Client({
   intents: [
     GatewayIntentBits.Guilds,
@@ -15,27 +20,25 @@ const client = new Client({
     GatewayIntentBits.GuildVoiceStates,
     GatewayIntentBits.DirectMessages,
     GatewayIntentBits.DirectMessageTyping,
-		GatewayIntentBits.MessageContent,
+    GatewayIntentBits.MessageContent,
   ],
   partials: [Partials.Message, Partials.Channel, Partials.Reaction],
 });
-require("dotenv").config({ path: path.resolve(__dirname, '.env'), quiet: true });
 
-// Commands Handler
+dotenv.config({ path: path.resolve(__dirname, ".env"), quiet: true });
 
-client.commands = new Collection();
+client.commands = new Collection<string, Command>();
 
 const commandsPath = path.join(__dirname, "commands");
 const commandFiles = fs
   .readdirSync(commandsPath)
-  .filter((file) => file.endsWith(".js"));
+  .filter((file) => file.endsWith(".ts"));
 
 for (const file of commandFiles) {
   const filePath = path.join(commandsPath, file);
-  const command = require(filePath);
-  // Set a new item in the Collection with the key as the command name and the value as the exported module
-  if ("data" in command && "execute" in command) {
-    client.commands.set(command.data.name, command);
+  const mod = await import(pathToFileURL(filePath).href);
+  if ("data" in mod && "execute" in mod) {
+    client.commands.set(mod.data.name, mod as Command);
   } else {
     console.log(
       `[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`
@@ -43,16 +46,15 @@ for (const file of commandFiles) {
   }
 }
 
-// Events Handler
-
 const eventsPath = path.join(__dirname, "events");
 const eventFiles = fs
   .readdirSync(eventsPath)
-  .filter((file) => file.endsWith(".js"));
+  .filter((file) => file.endsWith(".ts"));
 
 for (const file of eventFiles) {
   const filePath = path.join(eventsPath, file);
-  const event = require(filePath);
+  const mod = await import(pathToFileURL(filePath).href);
+  const event = mod as Event;
   if (event.once) {
     client.once(event.name, (...args) => event.execute(...args));
   } else {
