+++ package.json
@@ -3,10 +3,14 @@
   "version": "2.0.0",
   "description": "The Discord bot for r/realme Community Discord server.",
   "homepage": "https://github.com/raymond-1227/realmeme-superfart-rebooted",
-  "main": "index.js",
+  "type": "module",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "bun index.ts",
+    "deploy": "bun deploy.ts",
+    "undeploy": "bun undeploy.ts",
+    "typecheck": "tsc --noEmit"
   },
   "author": "Raymond Hsu",
   "license": "MIT",
@@ -21,5 +25,10 @@
     "libsodium-wrappers": "^0.8.4",
     "ms": "^2.1.3",
     "opusscript": "^0.1.1"
+  },
+  "devDependencies": {
+    "@types/ms": "^2.1.0",
+    "@types/node": "^22.0.0",
+    "typescript": "^5.7.0"
   }
 }
